// src/app/admin/user-admin-interface.ts
export interface IUserAdminInterface {

    Id: string;
    SID: string;
    FirstName: string;
    LastName: string;
    Timestamp:string;
    Telephone: string;
    EmailAddress: string;
    ManagerSID: string;
    HasSubordinates: boolean;
    RegionId: string;
    Role: string;

}